выносит функцию отрисовки элементов в модуль
