+++ src/main.js
@@ -1,3 +1,4 @@
+import {renderElement, RenderPosition} from './utils/render.js';
 import {createTripTabs} from './view/trip-tabs-view.js';
 import {createTripFilters} from './view/trip-filters-view.js';
 import {createTripSort} from './view/trip-sort-view.js';
@@ -8,23 +9,12 @@ import {createTripEvent} from './view/trip-event-view.js';
 import {createTripEventEditor} from './view/trip-event-editor-view.js';
 import {createTripStatistics} from './view/trip-statistics-view.js';
 
-const RenderPosition = {
-  BEFOREBEGIN: 'beforebegin',
-  AFTERBEGIN: 'afterbegin',
-  BEFOREEND: 'beforeend',
-  AFTEREND: 'afterend',
-};
-
 const TRIP_EVENTS_COUNTER = 3;
 
 const tripMainContainer = document.querySelector('.trip-main');
 const tripTabsContainer = document.querySelector('.trip-controls__navigation');
 const tripFiltersContainer = document.querySelector('.trip-controls__filters');
 const tripEventsContainer = document.querySelector('.trip-events');
-
-// statistics
-
-const renderElement = (container, element, position) => container.insertAdjacentHTML(position, element);
 
 renderElement(tripMainContainer, createTripInfo(), RenderPosition.AFTERBEGIN);
 renderElement(tripTabsContainer, createTripTabs(), RenderPosition.BEFOREEND);
