+++ src/utils/render.js
@@ -1,0 +1,10 @@
+const RenderPosition = {
+  BEFOREBEGIN: 'beforebegin',
+  AFTERBEGIN: 'afterbegin',
+  BEFOREEND: 'beforeend',
+  AFTEREND: 'afterend',
+};
+
+const renderElement = (container, element, position) => container.insertAdjacentHTML(position, element);
+
+export {renderElement, RenderPosition};
